Use `string replace` instead of `sed` (#87)
It seems like this was broken on my machine (OSX, 10.14.6) when using the completions. It would strip the `, ` but it would not add a proper newline. Instead it added nothing. So it just become a giant single line. Basically, it didn’t work.

With this change, the string gets split properly. Since this is fish shell, the `string replace` command is guaranteed to be there.