Add a CLI interface to edit the patches
`tldr --patch` to create/edit a new patch file.
`tldr --overwrite` to create/edit a new page file.

The logic of this is simply: use the $EDITOR variable to initiate an
editor on CUSTOM_PAGES_DIR/page.{patch,page}, exentension is determined
by which flag you pass.

It also checks that CUSTOM_PAGES_DIR exists and creates it if it does
not.