Implement `cache_dir` config file option
This required that the `Cache` always holds a `PathBuf` with its base
path and a cached `PathBuf` of the pages directory. This implies that
some cache operations (such as clearing) are now instance methods
instead of class methods, because the instance variables are used
(we always re-resolved the paths before).

On the way, I stumbled upon some oddities and added `FIXME`s. Please let
me know what you think about them :)