Fix clippy lints
This makes `Cache::list_pages` infallible and I am not 100% sure how I
fell about that. Before, it was impossible for it to fail, because the
existence of the path and the directory were already checked by some
other code (`Cache` construction and `check_cache` respectively), but
this makes the method unintuitive to use (it just returns an empty list
if there is nothing to traverse).