Upgrade to 2021 edition of Rust
To do so, we need to bump the MSRV to 1.56 (released 10/2021). While at
it, I actually bumped it all the way to 1.62, where the latest version,
(1.62.1) has been released on July 19, which is just more than one month
ago. Additionally, I removed `clippy.toml` and instead specified the
MSRV in `Cargo.toml`. This will also print an error when trying to
compile with an older compiler.