Add "random" command to fetch random tldr page
Added a new "--random" command that allows users to get a random page from tldr. Logics for picking a random page were implemented in the Cache class. The rand crate was also included to handle random selection of pages. This new feature gives users a way to discover new commands they may not be aware of.