Add random page display function
This commit adds the functionality to display a random page using -R or --random option. This new feature was implemented in bash, zsh and fish tealdeer by adding a new line or modifying existing ones. The changes also include calling a random function in the cache file and descriptive comments were added in the docs and src/cli.rs.